demo/python/requirements.txt
grpcio==1.29.0
grpcio-tools==1.29.0
protobuf==3.12.0
six==1.14.0
